+++ client/changelog.md
@@ -3,6 +3,7 @@
 - Added auto-focus on text field in "Volume" menu tab.
 - Fixed crash when switching servers on a proxy. ([#509](https://github.com/plasmoapp/plasmo-voice/issues/509))
 - Fixed percent rendering above player head breaks rendering on >1.21.1.
+- UDP connection is now considered established only after response from server is received.
 
 ### Server
 - Fixed "Incorrect behavior on resuming AudioSender". ([#501](https://github.com/plasmoapp/plasmo-voice/issues/501))